Refactor

+++ src/main/java/com/example/xbankbackend/models/BankAccount.java
@@ -4,14 +4,13 @@ import com.example.xbankbackend.enums.BankAccountType;
 import com.example.xbankbackend.enums.CurrencyType;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
 public class BankAccount {
     private UUID accountId;
     private UUID userId;
-    private BigDecimal balance;
+    private Float balance;
     private CurrencyType currency;
     private BankAccountType accountType;
 }
